airflow_sparking_flow/requirements.txt
apache-airflow 
apache-airflow-providers-apache-spark 
pyspark
apache-airflow-providers-openlineage>=1.8.0